# webserver/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/activity-logs/{id}:
    put:
      tags:
      - activity-log-controller
      operationId: updateEntry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityLog"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - activity-log-controller
      operationId: deleteEntry
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /api/sport-logs:
    get:
      tags:
      - sport-log-controller
      operationId: getAllEntries
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SportLog"
    post:
      tags:
      - sport-log-controller
      operationId: createEntry
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SportLog"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /api/sleep-entries:
    get:
      tags:
      - sleep-entry-controller
      operationId: getAllEntries_1
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepEntry"
    post:
      tags:
      - sleep-entry-controller
      operationId: createEntry_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SleepEntry"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/meetings:
    get:
      tags:
      - meeting-controller
      operationId: getAllEntries_2
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MeetingEntry"
    post:
      tags:
      - meeting-controller
      operationId: createEntry_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingEntry"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/gratitude:
    get:
      tags:
      - gratitude-controller
      operationId: getAllEntries_3
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GratitudeEntry"
    post:
      tags:
      - gratitude-controller
      operationId: createEntry_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GratitudeEntry"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GratitudeEntry"
  /api/food-logs:
    get:
      tags:
      - food-log-controller
      operationId: getAllEntries_4
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FoodLog"
    post:
      tags:
      - food-log-controller
      operationId: createEntry_4
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodLog"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/diary:
    get:
      tags:
      - diary-controller
      operationId: getAllEntries_5
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiaryEntry"
    post:
      tags:
      - diary-controller
      operationId: createEntry_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiaryEntry"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DiaryEntry"
...
  /api/activity-logs:
    get:
      tags:
      - activity-log-controller
      operationId: getAllEntries_6
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityLog"
    post:
      tags:
      - activity-log-controller
      operationId: createEntry_6
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityLog"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/sport-logs/{id}:
    delete:
      tags:
      - sport-log-controller
      operationId: deleteEntry_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/sleep-entries/{id}:
    delete:
      tags:
      - sleep-entry-controller
      operationId: deleteEntry_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/meetings/{id}:
    delete:
      tags:
      - meeting-controller
      operationId: deleteEntry_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /api/gratitude/{id}:
    delete:
      tags:
      - gratitude-controller
      operationId: deleteEntry_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/food-logs/{id}:
    delete:
      tags:
      - food-log-controller
      operationId: deleteEntry_5
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/diary/{id}:
    delete:
      tags:
      - diary-controller
      operationId: deleteEntry_6
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
...
    ActivityGroup:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    ActivityLog:
      type: object
      properties:
        id:
          type: string
        persons:
          type: string
        location:
          type: string
        startTime:
          type: string
          example: 14:30:00
        endTime:
          type: string
          example: 14:30:00
        date:
          type: string
          format: date
        activity:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/ActivityGroup"
          uniqueItems: true
        personTags:
          type: array
          items:
            $ref: "#/components/schemas/PersonTag"
          uniqueItems: true
        personList:
          type: array
          items:
            type: string
        durationMinutes:
          type: integer
          format: int32
    PersonTag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameLower:
          type: string
...
    SportLog:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        measurement:
          type: number
          format: double
        measurementUnit:
          type: string
        startTime:
          type: string
          example: 14:30:00
        endTime:
          type: string
          example: 14:30:00
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        notes:
          type: string
        user:
          $ref: "#/components/schemas/User"
        durationHours:
          type: number
          format: double
    SleepEntry:
      type: object
      properties:
        id:
          type: string
        fromTime:
          type: string
          example: 14:30:00
        untilTime:
          type: string
          example: 14:30:00
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        notes:
          type: string
        user:
          $ref: "#/components/schemas/User"
        hours:
          type: number
          format: double
...
    MeetingEntry:
      type: object
      properties:
        id:
          type: string
        place:
          type: string
        attendants:
          type: string
        startTime:
          type: string
          example: 14:30:00
        endTime:
          type: string
          example: 14:30:00
        date:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        durationMinutes:
          type: integer
          format: int32
...
    GratitudeEntry:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        eventDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
    FoodLog:
      type: object
      properties:
        id:
          type: string
        carbohydrates:
          type: number
          format: double
        kcal:
          type: integer
          format: int32
        dateTime:
          type: string
          format: date-time
        foodItems:
          type: string
        createdAt:
          type: string
          format: date-time
        notes:
          type: string
        user:
          $ref: "#/components/schemas/User"
        tags:
          type: array
          items:
            $ref: "#/components/schemas/FoodTag"
          uniqueItems: true
        foodItemList:
          type: array
          items:
            type: string
    FoodTag:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameLower:
          type: string
    DiaryEntry:
      type: object
      properties:
        id:
          type: string
        legacyDescription:
          type: string
        diaryReference:
          $ref: "#/components/schemas/DiaryReference"
        significance:
          type: string
          enum:
          - MINOR
          - NORMAL
          - MAJOR
        eventDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        startTime:
          type: string
          example: 14:30:00
        endTime:
          type: string
          example: 14:30:00
        user:
          $ref: "#/components/schemas/User"
        description:
          type: string
        points:
          type: integer
          format: int32
        durationMinutes:
          type: integer
          format: int32
    DiaryReference:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        descriptionLower:
          type: string
        user:
          $ref: "#/components/schemas/User"
...